TPA-1955: Lots of pretty big changes. One split out the various files so the SDK is easier to work with and easier to follow. Two add in various callbacks which will let the devs of the SDK access the data from the initial sdk request. Three add the ability to not render the choosePayer (step 3) step of the SDK. If this is set doneChoosePayer will provide the necessary function in order to make the SDK render the next form. Add everything to documentation.
