+++ assets/react/components/payer-info.jsx
@@ -1,0 +1,89 @@
+import React, { Component } from 'react';
+import Popup from 'react-popup';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faQuestionCircle } from '../util/font-awesome-icons';
+
+export class ControlledPopup extends Popup {
+  constructor(props) {
+    super(props);
+  }
+
+  componentDidUpdate() {
+    super.componentDidUpdate();
+    if (this.props.donePopUp) {
+      this.props.donePopUp();
+    }
+  }
+}
+
+export default class PayerInfo extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      popUpActive: false
+    };
+  }
+  render() {
+    const { payer, donePopUp } = this.props;
+    const { popUpActive } = this.state;
+    let message = `
+          Before you proceed, make sure you have registered on ${payer.website_home_url_netloc}
+          and have your ${payer.name} username and password at your fingertips.
+      `;
+    if (payer.has_security_questions) {
+      message += ` Also, have all ${payer.name} security questions and answers written down before you enroll below.`;
+    }
+    return (
+      <div>
+        <img src={payer.logo_url} style={{ maxWidth: '400px' }}></img>
+        <div style={{ display: 'flex' }}>
+          <h3>Enter Credentials for {payer.website_home_url_netloc}</h3>
+          <FontAwesomeIcon
+            icon={faQuestionCircle}
+            size="lg"
+            onClick={() => {
+              Popup.close();
+              Popup.create({
+                title: 'Help Getting Started',
+                content: `If you've not yet made an account with ${payer.website_home_url_netloc}, make one there first.`,
+                buttons: {
+                  left: [
+                    {
+                      text: 'Go To Payer Site!',
+                      action: () => {
+                        window.open(payer.register_url, '_blank');
+                      }
+                    }
+                  ],
+                  right: [
+                    {
+                      text: 'Ok!',
+                      action: () => {
+                        Popup.close();
+                        this.setState({ popUpActive: false });
+                      }
+                    }
+                  ]
+                }
+              });
+              this.setState({ popUpActive: true });
+            }}
+          />
+          <ControlledPopup
+            closeBtn={false}
+            donePopUp={popUpActive ? donePopUp : null}
+          />
+        </div>
+        <p
+          style={{
+            border: 'solid',
+            padding: '10px',
+            backgroundColor: '#FCF8E3'
+          }}
+        >
+          {message}
+        </p>
+      </div>
+    );
+  }
+}
