+++ assets/react/requests/validate-credentials.js
@@ -1,0 +1,17 @@
+import { sdkAxios } from '../services/axios';
+
+export const validateCredentials = async ({
+  taskId,
+  policyHolderId,
+  email
+}) => {
+  const validateCredsResponse = await sdkAxios.get(
+    `https://app.tpastream.com/sdk-api/validate-credentials/${policyHolderId}/${taskId}`,
+    {
+      params: {
+        email: email
+      }
+    }
+  );
+  return validateCredsResponse.data.data;
+};
